Consolidate VTK-XML parsing/encoding primitives into src/io/vtk
Mesh VTU, grid VTI, and ntree HTG each hand-rolled their own copies of the
same XML tag/region/attribute parsing, base64 codec, DataArray/Encoding
framing, and (un)compressed binary decode. Extracts all of it into
src/io/vtk/{read,write}, built on the existing src/io/deflate codec, and
updates all five read/write consumers to call the shared implementation.

Removes ~613 lines of duplicated parsing logic. Grid VTI and ntree HTG
read paths gain compressed-input support as a side effect of sharing the
Encoding/decode machinery with the VTU reader; a round-trip test for
compressed VTI was added to prove it. Compressed VTI/HTG write support is
not added here (would require an API-shape change like VTU's
Compressed/Uncompressed split, out of scope for this consolidation).
