+++ src/geometry/grid/write/vti/mod.rs
@@ -1,4 +1,7 @@
-use crate::{geometry::grid::Grid, io::NpyType};
+use crate::{
+    geometry::grid::Grid,
+    io::{NpyType, vtk::write::data_array},
+};
 use std::{
     array::from_fn,
     fs::File,
@@ -39,7 +42,7 @@ where
         file,
         "        <DataArray type=\"{}\" Name=\"data\" NumberOfComponents=\"1\" format=\"binary\">{}</DataArray>",
         vtk_type(T::DESCR),
-        payload(&data)
+        data_array(&data)
     )?;
     writeln!(file, "      </CellData>")?;
     writeln!(file, "    </Piece>")?;
@@ -62,34 +65,4 @@ fn vtk_type(descr: &str) -> &'static str {
         "<f8" => "Float64",
         _ => "UInt8",
     }
-}
-
-fn payload(data: &[u8]) -> String {
-    let mut buffer = Vec::with_capacity(8 + data.len());
-    buffer.extend_from_slice(&(data.len() as u64).to_le_bytes());
-    buffer.extend_from_slice(data);
-    base64(&buffer)
-}
-
-fn base64(bytes: &[u8]) -> String {
-    const ALPHABET: &[u8; 64] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    let mut out = String::with_capacity(bytes.len().div_ceil(3) * 4);
-    for chunk in bytes.chunks(3) {
-        let triple = ((chunk[0] as u32) << 16)
-            | ((*chunk.get(1).unwrap_or(&0) as u32) << 8)
-            | (*chunk.get(2).unwrap_or(&0) as u32);
-        out.push(ALPHABET[(triple >> 18 & 63) as usize] as char);
-        out.push(ALPHABET[(triple >> 12 & 63) as usize] as char);
-        out.push(if chunk.len() > 1 {
-            ALPHABET[(triple >> 6 & 63) as usize] as char
-        } else {
-            '='
-        });
-        out.push(if chunk.len() > 2 {
-            ALPHABET[(triple & 63) as usize] as char
-        } else {
-            '='
-        });
-    }
-    out
 }
