+++ src/io/vtk/mod.rs
@@ -1,0 +1,12 @@
+pub mod read;
+pub mod write;
+
+use std::io::{Error, ErrorKind};
+
+pub fn invalid(message: String) -> Error {
+    Error::new(ErrorKind::InvalidData, message)
+}
+
+pub fn unsupported(message: &str) -> Error {
+    Error::new(ErrorKind::Unsupported, message.to_string())
+}
